api/requirements.api.txt
python-dotenv==1.0.1
Flask==2.2.2
gunicorn==22.0.0
pika==1.3.2
Werkzeug==2.2.2